Sensor_fusion/requirements.txt
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.0
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
msgpack==1.1.0
numpy==2.0.2
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
pyproj==3.6.1
python-can==4.5.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.13.1
six==1.16.0
typing-extensions==4.12.2
wrapt==1.17.0
zipp==3.21.0
